use int64 for consistency

+++ pkg/apis/pxc/v1/pxc_backup_types.go
@@ -56,7 +56,7 @@ type PXCBackupSpec struct {
 	// RunningDeadlineSeconds is the number of seconds to wait for the backup to transition to the 'Running' state.
 	// Once this threshold is reached, the backup will be marked as failed.
 	// When unspecified, uses the value from the parent cluster's .spec.backup.runningDeadlineSeconds (which defaults to 5m).
-	RunningDeadlineSeconds *int32 `json:"runningDeadlineSeconds,omitempty"`
+	RunningDeadlineSeconds *int64 `json:"runningDeadlineSeconds,omitempty"`
 }
 
 type PXCBackupStatus struct {
